+++ docs/changes.rst
@@ -15,6 +15,7 @@ Pending
 * Toggle tracking the toolbar's queries when using
   ``debug_toolbar.store.DatabaseStore`` with ``SKIP_TOOLBAR_QUERIES``.
 * Fixed font family for code blocks and stack traces in the toolbar.
+* Added test to confirm Django's ``TestCase.assertNumQueries`` works.
 
 6.1.0 (2025-10-30)
 ------------------